pyproject.toml
[project]
name = "ticketmaster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.46",
    "celery>=5.6.2",
    "chromadb>=1.4.1",
    "fastapi>=0.128.4",
    "langchain-google-genai>=4.2.0",
    "langchain-ollama>=1.0.1",
    "pandas>=3.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "redis>=7.1.0",
    "sentence-transformers>=5.2.2",
    "sqlalchemy[asyncio]>=2.0.46",
    "streamlit>=1.19.0",
]
